Added HUD and inventory
Hud only consists of health and mana bars and the inventory bar which can be extended and retracted to show or hide the player's currently owned items.

Still need to enable functionality between player picking up items and them adding to the inventory as well as number overlays to show the quantity of each item the player owns.

Also modified hud_inv items to be separate from their inventory backgrounds to allow item icon transparency.