doorButtons are now interactable
doorButtons can be queried by and interacted with the player. When player activates them, it and it's sibling button across the doorway will do their animation together and their colliders will be moved away (for some reason destroy() gives an error). This functionality has been moved from the button class to the room class.
Still need to implement newRoom() functionality.