Enemies' sprites now load at startup
All enemy spriteSheets are loaded into memory when the game starts instead of when a room spawn during a playthrough. This is more performant and best coding practice.

Also re-written Enemies.lua to have a table of functions that spawn enemies, to more easily integrate into the random generating nature of the game.