updated nodes and enemies spawn rates
Both resource nodes and enemies have a 20% chance to not spawn at all. If they do spawn, enemies spawn in a group of 3 unless a demi boss spawns. Resource nodes also spawn in a group of 3 with a 50% chance to spawn an additional group of 3 in the same room.

Also updated the dungeon, player, and playState to spawn the boss room in the middle of the dungeon and a starting room in one of the 4 adjacent rooms to the boss room where the player also spawns in.