Camera now pans into newly spawned room
Implemented a new gamestate that handles panning to a newly spawned room before panning back to player. Player control is suspended during this camera movement.

Might need to tweak the timing of the pans later on.
May also want to experiment with different easing functions to make the camera pan look better.