Implemented fog of war
Each room now has a 'fog' covering any doorway that doesn't yet lead to another existing room. Once the player activates a doorButton and spawns another room, the fog gradually disappears. Still need to implement physical colliders to prevent players from walking into unspawned rooms.

Also tweaked the reveal speed of newly spawned room.

Also implemented a new timer which tweens the value of an object's property, however this is yet unused as it was initially intended to be used to lift the fog, but I decided to go with a different approach.