Updated craftingNodes.lua
Added art and animations for the furnace, grill, and spinning wheel. Still need to add sawmill as well as tweak all nodes' animations and colliders as well as add shadows, particles, etc.

Updated craftingNodes.lua to have an 'idle' and 'crafting' animation for each node as well as move some code from the object definition to the class definition.