Began crafting menu game state
Added placeholder assets for the enchanted crafting table and crafting menu.

Added crafting menu game state, however it only has basic functionality (opening and closing the menu)

Still need to improve the sprite for the crafting table and give it it's own particle system. Also need to continue implementing the functionality and look of the crafting menu. Finally, need to modify player.lua to sync up animations when the player crafts and obtains a new piece of armor.