Player can now craft all items from crafting table
Player is now able to craft all weapons, armor pieces, vials, and potions through the crafting menu brought up by the enchanted crafting table as long as they have all quantities of all items required to craft. Player inventory, HUD, and sprite will reflect any newly obtained or lost items as well.

Also rewritten player.lua to have all weapons and armor be included in the player's inventory (but not the HUD inventory).

Now all that's left is the design for the crafting menu itself and to improve the design of the enchanted crafting table.