updated resource nodes and boss room
Resource nodes now have a 1/6 chance of spawning a set of arcane shards along with their usual resource items.

The boss room will no longer contain the boss itself, but rather just a ladder that will take the player to the boss. This was done for (1) smoother gameplay; all other entities will be cleared from the game, (2) so that the arena can be fully seen from the player, with no vertical walls covering any part of the floor which is important for the player's ability to react and avoid all attacks, and (3) so the boos room can be made a different size if need be.