Implemented dialog system
Updated the font to include lowercase letters and some special characters.

Implemented the dialog system for the player character that will let the player know information regarding what items they need when they fail a craft at any of the crafting nodes.
(Will probably have to find another place to draw the player dialog, as it is buggy if the player itself draws it. It's now just drawn in the playstate, but will probably move to a 'user interface' file.

Also fixed a bug causing vines to never spawn arcane shards.