Updated crafting table sprite
crafting table now has a tome containing knowledge of how to use the magic imbued in arcane shards to craft weapons and armor from components as well as a mortar and pestle along with some vials which are used to create potions.

Also updated the font, changed some dialogs, and created some testing functionality to the ladder.