action buttons now visually depress
action button blank now somewhat animate between an unpressed and pressed sprite so when the player presses a button, the action button visually pressed as well. Only minor issue is that when the player opens a new room, the new state stops input from going into the play state where the action button updates, and so the visual effect stops until the camPanState is over.

Also started enemies on random animation frames when they spawn.