Updated camera to have damping
Updated usage of the camera to use lockPosition() instead of lookAt() which allows the camera to take care of easing functions while all I need to do is set the camTarget to whatever I want the camera to pan to. This also creates a rubber band effect where the camera will follow the player with more realistic inertia instead of remaining perfectly snapped to the player's position at all times.
This will help prepare for the camera movement in combat.