Updated TAP and HOLD button functionality
Updated actionButtons' tapping and holding functionality to work properly and intuitively, as it was not before

Previously, holding the button would be recognized only after the button is released, making it more of a short press/long press situation. Now, holding the button triggers after the button is held for a certain amount of time, without the player needing to release the button.