Implemented advanced enemy targeting system
Player now has a 'history' of targeted enemies that can be used to switch between them in a smart way. Currently, the player can remember the last 6 enemies they targeted, but this can change in the future.

Currently, I made it such that when a player disengages from combat, their 'history' of targeted enemies is cleared and will build from scratch when they start switching targets. I might go back and try to implement this feature after playtesting if it seem like it should be added.