Targeting system bug fix
Fixed a bug caused by trying to switch targets when in combat and only one enemy is nearby as well as when the player is in combat and searches for nearby targets and none can be found (targeted enemy is beyond query area but still not yet at the disengage distance). This resulted in the currentEnemy property being nil when it wasn't supposed to. A few conditional statements was all that was needed.