Created projectile file
Player will now launch projectiles at the enemy they're targeting with a cooldown of 1s (might adjust later) with a constant speed. The only projectile added so far is the stone which will be thrown when the player has no bow or staff yet (the default attack). The player also keeps track of the line of sight between itself and the enemy target, to know when to not attack.

Also modified entities.lua to use a separate table to hold all entities that are within draw distance, reducing the total number of math.abs() calls.

Lastly, modified the room walls in rooms.lua to be a part of another collision class which will be used to check player-enemy line of sights.