Projectiles now home in and angle to target
Projectiles  now home in while still traveling at a constant speed as well as angle their sprites toward the target, which is very noticeable in the arrow projectile. This was done by calculating the angle from the projectile's position to the target's position.

Weapon animations are only still very basic, will have to sync up animations to attacking. Also will try to angle the weapon toward the target.