Projectiles now hover above ground
To be consistent with all other entities, projectile's actual position is lower than their sprite make them look. This is in order to achieve the somewhat 3d effect.

Also, because projectiles launched from the staves are higher, their sprites will be lowered to the ground over the course of their travel, so as to look like they impact even the smallest enemies.

Projectiles are now entities once again.