Fixed weapon sprite clipping bugs
Updated dungeon background and foreground sprites, room lights, and inner walls in order to prevent clipping from tall weapons such as the staff. Also updated the player to turn the other way when the player has a bow or staff equipped and if right in front of a wall, facing it in order to prevent clipping from horizontal weapon sprites like the bow.

Also fixed a projectile bug where it would launch the enemy in the wrong direction upon landing. This was because it was calculating its velocity for the next frame before checking for a collision.