Player weapon sprites now have collision
The player's collider now has two rectangle shapes on either side fixed to it which can produce collisions when it needs to (such as when a weapon sprite is drawn). This will prevent clipping whenever a weapon is drawn to the screen by essentially giving the player a larger collision box to accommodate the sprite. This can also be extended to be used for consumables sprites as well as protection spells' sprites.