Implemented protection magics floating symbols
Protection magics are fully implemented (other than preventing damage, which is yet to be introduced into the game) and the player can now activate, deactivate, and switch between physical and magical protection.
When a protection magic is activated, four symbols of the appropriate icon float around the player and have their own collision bow to prevent clipping.