Tweaked melee enemies' attacks
Melee enemies now all have the same mass, reduced linear damping compared to other enemies and the player, and lunge toward the player at a slightly slower speed. This makes it more clear that the enemies are actually lunging toward the player, however this also means that enemies have a slight chance of completely missing their lunge if the player is at the right distance, speed, angle, etc. which I'm actually alright with because it's very difficult to dodge enemies like this consistently, even with plenty of room and only one enemy to deal with.

Also, the player can now successfully block damage of both melee and projectile based physical and magical attacks. Knockback of melee attacks can't be completely removed as it is one physical collider running into the player's, but it is greatly reduced when using physical protection magics. I might decide to have make the protection from projectiles greatly reduce knockback in the same way as opposed to completely ignoring any knockback effects when protecting correctly, just to make everything more consistent.