Mana now drains while protection magics are active
When using a protection spell, mana drains every 0.5s. The magnitude of this drain is affected by armor (higher tier armor means a slower drain rate)

When mana is depleted, protection magics are deactivated. If the player tries to reactivate without first restoring mana, a random dialog will be said by the player about not having enough mana.

Also tweaked enemies diagonal movement speed to be slightly higher as before it was an noticeable and unnatural decrease from straight-line speed to diagonal speed.