Movement is now done with forces
Setting the velocity of a collider is not framerate independent. Nor is it a good approach to begin with as entities will ignore the impulses (knockback) of projectiles.

Movement is now overhauled not only to use angles in order to normalize vectors, it now uses forces to accelerate and decelerate colliders. Now, movement is consistent regardless direction or framerate and now feels great!

Also, as it wasn't stated in the previous commit, tier 2 and 3 demons take reduced magical damage, while tier 2 and 3 orc take reduced physical damage.