Implemented tornado special attack
Created specialAttacks file which will handle the creation of the special attacks for the demi bosses and the final boss.

Added the tornado special attack for the orc_t3 which will launch two tornados in quick succession and, after travelling in the direction of the angle perpendicular from the orc to the player, will begin travelling directly toward the player, dealing 'pure' damage and knocking back the player upon contact (with a 0.5s damage cooldown).

Will probably need to tweak the behavior of the tornado (movement speed, knockback, damage, oscillation rates, etc. later)