restructured room floors
Floor now are separated from a room's background and foreground sprites so that floor effects can be added into the game. These effects are drawn everything except for the floor itself so as to appear 'beneath' all entities and dungeon walls.

Also added the fire insignia which is the first half the the demon_t3's special attack. The insignia which shows the player where some damaging flames will appear is a floor effect, but the flames that will spawn on top of the insignia will be entities.