Dialog systems can now display damage numbers
Dialog systems can now show the damage an entity takes along with a color to represent the damage type. Gray is for physical, blue is for magical, and red is for pure (which can't be protected against and it used for special attacks)

The fire circle special attack has been modified to use a single, large collider to damage the player instead of the multiple colliders used to move the particle systems. This make taking damage and knockback much more consistent at a steady rate of 0.1s per damage tick. (As opposed to running across the fire circle, taking a large burst of damage from the multiple damage sources, then waiting 0.5s for the next large burst)

Also, mushrooms can now also be dropped by enemies instead of fish (though they will roll to choose which one to drop) this should make getting potions mush easier and will make the fishing holes less pointless.