Added boss room phase to Playstate
Interacting with the ladder leading to the boss room now changes the 'phase' of the playstate the player is in from 'dungeon phase' to 'boss battle'. All entities and colliders that were used during the dungeon phase is removed at this time, and the player is put into a boss room. Still need improve the look of the boss room as well as implement a better transition from dungeon to boss chamber

Also added the aseprite file for the boss; a magma golem. Still need to add many improvement as well as animations for physical attacks, magical attacks, etc.