Implemented Boss' basic physical attack
Boss can now perform its basic physical attack which to slam the ground in front of it, sending a fissure toward the player.
Still need to finish the animation, and may also improve the fissure itself later on.

Also modified the combat action button to have varying distances where combat can be engaged/will be disengaged. This is so that the player can always target the boss during the final fight and so that the player won't disengage with the boss (the boss room is much larger than a regular dungeon room).