Added animated lava and improved flame tornados
Lava surrounding the arena is now animated as tiles, though I'm still considering improving or otherwise changing it. This animated lava does not show on the arena when a floor tile is filled with lava.

Flame tornados now consist virtually entirely of particles (much like the fireball) (there's still a shadow) and are more like flame pillars. They also no longer move in a circle like the tornados but otherwise function exactly the same way. They'll travel in a given angle for 2s, then begin travelling to the player. I'm considering having their movespeed increase as the boss loses health, as it is pretty easy to avoid them even with four flame tornados and a hard floor lava pattern.