Player can now 'drop' into a room
To enter a room, the player model drops down into it. This is currently used when the player enters the dungeon as well as the boss room.
 Still plan to create a particle effect for when the player lands, but for now, an animation frame and a shadow that scales with the player's 'distance' from the ground is pretty good.

I was planning to have an accompanying 'exit' player transition where the player would drop down into the ladder hole in order to exit from the current room, but I've decided to save that for later. For now, just a fade to black is good enough.

Also modified clock.lua to have the full transition when time runs out: fade to black, change to boss room, fade in, then drop player into the room.