Added game end screen
Implemented a very simple endscreen for when the player beats the game or dies. Currently they can choose to restart, but once the title screen is implemented, they can also choose to go to it from the endscreen.

The endscreen for the game over simply shows a large 'game over' text. This screen might be removed in the final game.
The endscreen for winning the game shows the player their completion time for the dungeon, boss, and total.

Also changed how the clock works. It now works as any other computer science major would expect; using division and modulus.

Also modified how the menu buttons (accept and decline) work. They have been removed from the creaftingMenuState and are now included with the rest of the action buttons. This makes them more portable and so they are able to be used in the endscreen and will also be used in the title screen once implemented.

Other changes include fixes such as: removing player projectiles that are active as the player dies, boss arena environmental hazards not damaging the player after the boss has been killed, being able to specify the target alpha that fadeState will fade to (allowing partial fade outs), player enter transition being screen resolution independent.