Updated flamePillar behavoir
flamePillar behavior has been reworked. Instead of targeting the player directly, each pillar now targets one of 5 points around the player. These points are arranged in the shape of a star. Every 0.5-1.5 seconds, the targeted point changes to the point across the player. Essentially, the flame pillars 'draw' the shape of a star, damaging the player as they change from one point to the next. This results in more engaging behavior when the player is running from the pillars as they spread out more and often cut off the player, making it much more difficult for the player to simply kite around all pillars in a small circle.