Camera pans to boss when it dies
Camera now pans to the boss to show the player it's death animation (not yet implemented, instead just showing it explode into particles), then camera pans back to player as the endscreen pops up.

Modified the flame pillar to change it's update function upon dying to fix a bug causing the flame pillar to still be able to damage the player during the time between the boss dying (health reaches 0) and it being removed from the game.