Fixed bug during boss' death animation
Fixed a bug where the player can still take damage from BossRoom's perimeter lava while the boss is dead because they can still move around while the camera pans to the boss to show the player the boss' death explosion.
Player still can move around in that time (might create a new gamestate in the future to prevent this) but the perimeter lava no longer damages the player once the boss is dead (or if the boss is nil or if the player is dead) much in accordance with the behavior of the rest of the environmental hazards of the BossRoom.

Also updated boss' updateHealth() function to let it die as soon as it's health reaches 0 (changed from making it die once the healthbar above it is completely depleted) This prevents another bug where the camera pans to the boss, then back to the player because boss stopped being a combat target.