Changed actionButtons' input mapping
Instead of numpad buttons, the actionButtons are now controlled by the arrow keys. This is just for the beta because keys are not yet remappable and not everyone has a keyboard with a numpad, but everyone should have arrow keys.

Also potentially fixed a bug where the game would crash with no error messages logged as soon as the boss battle started. I believe this was due to the boss spawning after the entitiesTable is cleared. The time the entitiesTable being cleared and the boss spawning was supposed to be 1 frame, however it might have just been long enough for 1 frame on a 60fps machine, but too short for 1 frame on a 144fps machine. Thus I increased the amount of time between clearing the table and spawning the boss so hopefully this major issue is resolved.

Also potentially fixed another bug where the boss has 0 current health but won't die. May have something to do with the fact that the boss takes 1/5 of displayed damage, and is sometimes left with a health value so small that it no longer has enough health to 'live' but boss.health.current==0 is false, so it doesn't die either. Changed the check to boss.health.current<=0.5, hopefully it fixes.

Also removed the console from opening, since this shouldn't happen for the beta build