Browser version beta release build
Browser beta version is ready to go!
Made some minor changes for the release: (1) changed the conf.lua file to revert resolution to 1200x900 because love.js can fullscreen the game extrernally. Still need to double check that this scales the game down on displays that are smaller than 1200x900. (2) returned the t.version as this was only for the love web builder which has been deprecated and doesn't work with windfield. I'm instead using Davidobot's updated fork of love.js which supports the latest version of love2d. (3) removed the quick exit from both the endscreen and the 'escape' keyboard shortcut as doing this in the browser simply freezes the game which is no good. Now in the endscreen, both 'PLAY AGAIN' and 'QUIT GAME' simply restart the game. The escape key is also used to exit browser fullscreen.