Changed HTML page for compatibility
Modified index.html to remove header so that the game takes the full resolution of the canvas, put the title and author in the loading screen, and changed the 'fullscreen' button to be in the center of the canvas and have white text.

Also added the itch.io ready zip package

Also added keyboard controls graphic, both regular and 2x scaled sizes which will be added to the itch.io webpage.