player feedback improvements
Names of all interactables (crafting nodes, resource nodes, door buttons, and ladder) now pop up above their sprites as UI elements in order to show the player that they're important.

Added a white font which is used for the UI elements text

Added "accept" and "exit" to the crafting menu screen next to their respective buttons to hopefully make them more obvious as to how to use the buttons during crafting.

When the player takes damage while using the crafting menu, the menu will be closed.

As player falls into the dungeon at the start of the game, the playstate uses a draw function slightly modified from the drawDungeonPhase function in order to prevent clipping. Now, all entities are drawn last during this transition so that (1) the player will be drawn above the fog of war, and (2) the crafting nodes can still be drawn 'in front of' the player.