added functional main menu to title screen
players can now navigate the main menu which is the first menu presented in the title screen. Additional menus that will be included are the tutorial menu and the settings menu, along with their sub-menus.

Players can currently choose to start the game or close the game. Currently the tutorial menu can be opened, but it just shows the controls guide.

Game can now safely close for both native and browser versions.

Added and modified some sounds including menu opening and closing as well as moving menu cursors, improved fish consume sfx.