Implemented saving capability with safeStorage
safeStorage, in combination with knife.serialize is responsible for reading and writing the settings file and ensuring that the settings file is always valid and safe to load as a script, no matter how the user tampers with the file itself. See safeStorage.lua for more details.

added the ability to save the fullscreen setting in the titleScreenState. I'll add the ability to save settings for everything else next.

Modified love.load to take in a table or arguments. One of which is ['keepSettings'] which will tell the game to not read and apply the settings from the settings file. This is used when the play dies or beats the game, then returns to the main menu. If this were not there, the game could reapply fullscreen settings and cause unnecessary visual issues.