Controller support added
Controller support is now fully added.

Controls have been overhauled to be able to use multiple keys/buttons/axes to switch on a single input. Keyboard users can now use arrow keys and/or numpad keys for action buttons. Controller users can use the left analog stick and/or dpad to control movement and navigation. You can even use both keyboard and controller inputs together if you really wanted.

Controls now listens for input presses (inputs that are down for the first frame only). This was done only for menu navigation in the titleScreen and craftingMenu as it felt really unnatural (especially when using a controller) to have navigation done on input release.

Removed the ability to re-map controls even from the settings file. I can return this functionality later if it's requested, I'd just need to implement validation on mappings. However I don't think it's worth the effort right now until people actually request it.

Fixed a bug where the craftinMenu didn't reset it's crafting options when the player restarted the game straight from dying or beating the game (as opposed to returning to the title screen).