Removed controller support, going back to 11.3
I've decided that controller support is simply not worth the amount of effort currently. May add it later if people want, but for now, just keyboard support.

Love2D doesn't really do the concept of a 'bottom/top/left/right' face button, instead it tries to make the A button on an xbox controller be the A button on a nintendo controller, even though they are in different positions. This makes it very difficult to implement a control scheme for my game that can work for majority of common controllers.
Furthermore, love.js seems to have a bug where my controller is recognized as a different controller on the web version, one with completely different mappings than on native. Will probably bring this up as an issue on davidobot's repo.
With all of this in mind, the best approach would be to have remappable controls and have the player setup their controller to work with the game. This would also require further validation to the settings file, or a complete encryption system to ensure users don't break the game by directly modifying the file.
All in all, it's simply too much work for something that probably won't be wanted by most. And for the few who do want it, they can just use a controller to key mapping program for now.

Next I need to move on to creating the tutorial.
I also moved back to Love2D version 11.3 as this is what's compatible with love.js and the only reason I updated to 11.4 was to include controller support.