Tutorial completed
Full tutorial is now in the game.

Modifications to the player and enemies were made to allow checking certain objectives for a few tutorials. These are detailed below:

Player now keeps track of if they successfully blocked damage.
Creating a new enemy returns the enemy in addition to adding it to the entitiesTable.
Projectiles now destroy themselves if the player is invincible.
ResourceNodes now return their resource nodes upon creating in addition to adding themselves to the entitiesTable.
Entities now only destroys colliders if they aren't already destroyed.

Fonts were modified to add symbols: "/()"

Each individual tutorial is now created in the TutorialStates' load() function. This was done because lua's require() tried to save memory by not re-loading things it already see are in memory. This caused a bug where, if the player tried to restart the tutorial, individual tutorials could still thing their objectives are complete, which made the tutorial broken.